move: redis 디렉토리로 redis 관련 클래스들 이동
