+++ src/main/java/com/newzet/api/common/cache/redis/RedisServerException.java
@@ -1,6 +1,4 @@
-package com.newzet.api.common.cache.exception;
-
-import org.springframework.data.redis.connection.RedisServer;
+package com.newzet.api.common.cache.redis;
 
 public class RedisServerException extends RuntimeException {
 	public RedisServerException() {
